requirements.txt
#robotframework
#robotframework-seleniumlibrary
#robotframework-appiumlibrary
#robotframework-cucumberjsonlibrary
behave
selenium
appium-python-client
#pytest-bdd
pabot
pyyaml